feat(prepare): add gemFileDir config option

+++ src/publish.js
@@ -2,7 +2,7 @@ const { unlink } = require('fs').promises;
 const execa = require('execa');
 
 module.exports = async function publish(
-  { gemHost, gemPublish = true },
+  { gemHost, gemPublish = true, gemFileDir = false },
   { cwd, env, logger, nextRelease: { version }, stdout, stderr },
   { gemFile, gemName, credentialsFile },
 ) {
@@ -22,5 +22,7 @@ module.exports = async function publish(
     logger.log(`Skip publishing to gem server because gemPublish is ${gemPublish !== false}`);
   }
 
-  await unlink(gemFile);
+  if (gemFileDir === false) {
+    await unlink(gemFile);
+  }
 };
